Fix tmva windows warning about 32/64 bit conversion
Just use from beginning 64 bit values

+++ tmva/tmva/src/Factory.cxx
@@ -2248,7 +2248,7 @@ TH1F* TMVA::Factory::EvaluateImportanceAll(DataLoader *loader, Types::EMVA theMe
   {
     SROC=ROC[x];
     for (uint32_t i = 0; i < VIBITS; ++i) {
-      if (x & (1 << i)) {
+      if (x & (uint64_t(1) << i)) {
    y = x & ~(1 << i);
    std::bitset<VIBITS>  ybitset(y);
    //need at least one variable
@@ -2338,7 +2338,7 @@ TH1F* TMVA::Factory::EvaluateImportanceShort(DataLoader *loader, Types::EMVA the
 
   for (uint32_t i = 0; i < VIBITS; ++i) {
     if (x & (1 << i)) {
-      y = x & ~(1 << i);
+      y = x & ~(uint64_t(1) << i);
       std::bitset<VIBITS>  ybitset(y);
       //need at least one variable
       //NOTE: if sub-seed is zero then is the special case
@@ -2446,7 +2446,7 @@ TH1F* TMVA::Factory::EvaluateImportanceRandom(DataLoader *loader, UInt_t nseeds,
       //removing global result because it is requiring a lot of RAM for all seeds
 
       for (uint32_t i = 0; i < 32; ++i) {
-         if (x & (1 << i)) {
+         if (x & (uint64_t(1) << i)) {
             y = x & ~(1 << i);
             std::bitset<32>  ybitset(y);
             //need at least one variable
